may_14/src/requirements.txt
chromadb
pypdf2
tiktoken
click
python-dotenv